+++ lib/service/config.dart
@@ -329,6 +329,12 @@ class Config with Service, NetworkAuthProvider, NotificationsListener {
     return _packageInfo?.version;
   }
 
+  String? get appStoreId {
+    String? appStoreUrl = MapPathKey.entry(Config().upgradeInfo, 'url.ios');
+    Uri? uri = (appStoreUrl != null) ? Uri.tryParse(appStoreUrl) : null;
+    return ((uri != null) && uri.pathSegments.isNotEmpty) ? uri.pathSegments.last : null;
+  }
+
 
   // Getters: Config Asset Acknowledgement